Create Wireless LAN Simulation with ns-3
This lab course will serve as an introduction to the field of 802.11 wifi simulations. The network simulator ns-3 will be used to illustrate certain characteristics of this
protocol family. Different WLAN techniques and topologies will be simulated to give an overview about this wireless network type. The network simulator ns-3 is a discrete-event network simulator. It is an open source project for simulating network environments, both cable-connected and wireless. The project is licensed under the GNU general public license version 2 and is available for use and participation at [1]. Unlike its predecessor ns-2, which was written in a mixture of C/C++ and the scripting language OTcl, ns-3 is completely implemented in object oriented C++. This enables high-performance simulation as well as easy code extension and reuse. The scripting language Python is used for the integrated build system, which handles compilation and linking as well as passing command-line arguments.