OMNeT++ Lab Part 1 and Part 2
OMNeT++ [1] is a discrete event based simulator. It can be used to simulate almost everything, such as queues, multi-processor-based systems, various hardware architectures, and business processes. For every area, there is a particular set of module compilations. OMNeT++ itself only provides a simulation framework. While OMNeT++ modules and models are built in the C++ programming language, the description of the (network) structure is written in the NED language (NEtwork
Description language). The latter can be done either in advance or during the simulation run time. Every object or model has distinct files.